Create restore-script-for-sqlcmd.sql
here's another stab at the script, using a standalone sqlcmd script. Programming in transact-sql is rather painfully reminiscent of COBOL without the modern features and bad string handling, but it has the virtue of portability, all you need is SQLCMD - or a SQLCMD-like client that understands some SQLCMD parsing features.